evals/requirements.txt
# Dependencies for generate_ai_responses.py
boto3>=1.34.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
